+++ src/main/java/com/example/roommate/applicationServices/LoginApplicationService.java
@@ -1,7 +1,7 @@
 package com.example.roommate.applicationServices;
 
 import com.example.roommate.annotations.ApplicationService;
-import com.example.roommate.dtos.forms.LoginForm;
+import com.example.roommate.values.forms.LoginForm;
 
 @ApplicationService
 public class LoginApplicationService {
