+++ src/main/java/com/example/roommate/controller/LoginController.java
@@ -1,6 +1,6 @@
 package com.example.roommate.controller;
 
-import com.example.roommate.dtos.forms.LoginForm;
+import com.example.roommate.values.forms.LoginForm;
 import com.example.roommate.applicationServices.LoginApplicationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
