+++ src/main/java/com/example/roommate/values/forms/LoginForm.java
@@ -1,4 +1,4 @@
-package com.example.roommate.dtos.forms;
+package com.example.roommate.values.forms;
 
 public record LoginForm(String username, String password) {
 }
