+++ src/test/java/com/example/roommate/tests/controller/LoginControllerTest.java
@@ -2,7 +2,7 @@ package com.example.roommate.tests.controller;
 
 import com.example.roommate.annotations.TestClass;
 import com.example.roommate.controller.LoginController;
-import com.example.roommate.dtos.forms.LoginForm;
+import com.example.roommate.values.forms.LoginForm;
 import com.example.roommate.applicationServices.LoginApplicationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
