+++ src/test/java/com/example/roommate/tests/controller/booking/PostBookTest.java
@@ -3,7 +3,7 @@ package com.example.roommate.tests.controller.booking;
 
 import com.example.roommate.annotations.TestClass;
 import com.example.roommate.exceptions.domainService.GeneralDomainException;
-import com.example.roommate.dtos.forms.BookDataForm;
+import com.example.roommate.values.forms.BookDataForm;
 import com.example.roommate.applicationServices.BookingApplicationService;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
