+++ src/test/java/com/example/roommate/tests/domain/ValuesTest.java
@@ -1,9 +1,9 @@
 package com.example.roommate.tests.domain;
 
 import com.example.roommate.annotations.TestClass;
-import com.example.roommate.dtos.forms.BookDataForm;
+import com.example.roommate.values.forms.BookDataForm;
 import com.example.roommate.values.domain.ItemName;
-import com.example.roommate.dtos.forms.LoginForm;
+import com.example.roommate.values.forms.LoginForm;
 import com.example.roommate.factories.ValuesFactory;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
